requirements.txt
curl-cffi>=0.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
pyyaml>=6.0

